% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAMatrix.R
\name{LAMatrix_main}
\alias{LAMatrix_main}
\title{Efficient eQTL Mapping with Local Ancestry}
\usage{
LAMatrix_main(snps, gene, cvrt = SlicedData$new(), local = SlicedData$new(),
  output_file_name = "", pvOutputThreshold = 1e-05,
  useModel = modelLINEAR, errorCovariance = numeric(), verbose = TRUE,
  output_file_name.cis = "", pvOutputThreshold.cis = 0, snpspos = NULL,
  genepos = NULL, cisDist = 1e+06, pvalue.hist = FALSE,
  min.pv.by.genesnp = FALSE, noFDRsaveMemory = FALSE)
}
\arguments{
\item{snps}{\code{SlicedData} object with genotype information.}

\item{gene}{\code{SlicedData} object with gene expression information.
Must have columns matching those of \code{snps}.}

\item{cvrt}{\code{SlicedData} object with additional covariates.
Can be an empty \code{SlicedData} object in case of no covariates.
The order of columns must match those in \code{snps} and \code{gene}.}

\item{local}{\code{SlicedData} object with local ancestry information.
The order of columns must match those in \code{snps}, \code{gene} and \code{cvrt}.
The order of rows must match those in \code{snps}.
Can be an empty \code{SlicedData} object and the analysis will be the same as \code{modelLINEAR}.}

\item{output_file_name}{\code{character}, \code{connection}, or \code{NULL}.
If not \code{NULL}, significant associations are saved to this file (all significant associations
if \code{pvOutputThreshold=0} or only distant if \code{pvOutputThreshold>0}).
If the file with this name exists, it is overwritten.}

\item{pvOutputThreshold}{\code{numeric}. Significance threshold for all/distant tests.}

\item{useModel}{\code{integer}. Either \code{modelLINEAR}, \code{modelANOVA},
\code{modelLINEAR_CROSS}, or \code{modelLOCAL}.
 \enumerate{
 \item Set \code{useModel = modelLINEAR} to model
 the effect of the genotype as additive linear and
 test for its significance using t-statistic.
 \item Set \code{useModel = modelANOVA} to treat genotype
 as a categorical variables and use ANOVA model and
 test for its significance using F-test.
 \item Set \code{useModel = modelLINEAR_CROSS} to add
 a new term to the model equal to the product of genotype and
 the last covariate; the significance of this term is
 then tested using t-statistic.
 \item Set \code{useModel = modelLOCAL} to model the
 effect of the genotype as additive linear while adjusting for a
 SNP-specific covariate (e.g. local ancestry) and test for genotype
 significance using t-statistic.
 }}

\item{errorCovariance}{{\code{numeric}. The error covariance matrix.
Use \code{numeric()} for homoskedastic independent errors. }}

\item{verbose}{{\code{logical}. Set to \code{TRUE} to display
more detailed report about the progress.}}

\item{output_file_name.cis}{\code{character}, \code{connection}, or \code{NULL}.
If not \code{NULL}, significant local associations are saved to this file.
If the file with this name exists, it is overwritten.}

\item{pvOutputThreshold.cis}{\code{numeric}. Same as \code{pvOutputThreshold}, but for local eQTLs.}

\item{snpspos}{{\code{data.frame} object with information about SNP locations,
must have 3 columns - SNP name, chromosome, and position.}}

\item{genepos}{{\code{data.frame} with information about transcript locations,
must have 4 columns - the name, chromosome, and positions of the left and right ends.}}

\item{cisDist}{{\code{numeric}. SNP-gene pairs within this distance are
considered local. The distance is measured from the
nearest end of the gene. SNPs within a gene are always considered local.}}

\item{pvalue.hist}{{\code{logical}, \code{numerical}, or \code{"qqplot"}.
Defines whether and how the distribution of p-values is recorded
in the returned object.}}

\item{min.pv.by.genesnp}{{
\code{logical}. Set \code{min.pv.by.genesnp = TRUE} to record
the minimum p-value for each SNP and each gene in the returned object.}}

\item{noFDRsaveMemory}{{
\code{logical}. Set \code{noFDRsaveMemory = TRUE} to save
significant gene-SNP pairs directly to the output files.}}
}
\value{
{
The detected eQTLs are saved in \code{output_file_name}
and/or \code{output_file_name.cis} if they are not \code{NULL}.
The method also returns a list with a summary of the performed analysis.
\item{param}{
Keeps all input parameters and also records
the number of degrees of freedom for the full model.}
\item{time.in.sec}{
Time difference between the start and
the end of the analysis (in seconds).}
\item{all}{
Information about all detected eQTLs.}
\item{cis}{
Information about detected local eQTLs.}
\item{trans}{
Information about detected distant eQTLs.}
The elements \code{all}, \code{cis}, and \code{trans}
may contain the following components
\describe{
\item{\code{ntests}}{
Total number of tests performed. This is used for FDR calculation.
}
\item{\code{eqtls}}{
Data frame with recorded significant associations.
Not available if \code{noFDRsaveMemory=FALSE}
}
\item{\code{neqtls}}{
Number of significant associations recorded.
}
\item{\code{hist.bins}}{
Histogram bins used for recording p-value distribution.
See \code{pvalue.hist} parameter.}
\item{\code{hist.counts}}{
Number of p-value that fell in each histogram bin.
See \code{pvalue.hist} parameter.
}
\item{\code{min.pv.snps}}{
Vector with the best p-value for each SNP.
See \code{min.pv.by.genesnp} parameter.
}
\item{\code{min.pv.gene}}{
Vector with the best p-value for each gene.
See \code{min.pv.by.genesnp} parameter.
}
}
}
}
\description{
\code{LAMatrix_main} is used to perform eQTL mapping with local ancestry when specifing \code{useModel = modelLOCAL}.
The eQTL mapping is based on the linear model assuming additive effect of genotype on gene expression.

This function performs the same analysis as MatrixEQTL::\link[MatrixEQTL]{Matrix_eQTL_main}
when use \code{useModel = modelLINEAR, modelANOVA, } and \code{modelLINEAR_CROSS}.
}
\examples{
library("LAMatrix")

n = 200;# Number of columns (samples)
nc = 10;# Number ofs covariates

cvrt.mat = 2 + matrix(rnorm(n*nc), ncol = nc);# Generate the covariates
snps.mat = floor(runif(n, min = 0, max = 3)); # Generate the genotype
local.mat = floor(runif(n, min = 0, max = 3)); # Generate the local ancestry

# Generate the expression vector
gene.mat = cvrt.mat \%*\% rnorm(nc) + rnorm(n) + 0.5 * snps.mat + 1 + 0.3 *local.mat;


#Create SlicedData objects
snps = SlicedData$new( matrix( snps.mat, nrow = 1 ) );
genes = SlicedData$new( matrix( gene.mat, nrow = 1 ) );
cvrts = SlicedData$new( t(cvrt.mat) );
locals = SlicedData$new( matrix(local.mat,nrow=1))

# Produce no output files
filename = NULL; # tempfile()

me = LAMatrix_main(
 snps = snps,
 gene = genes,
 cvrt = cvrts,
 local = locals,
 output_file_name = filename,
 pvOutputThreshold = 1,
 useModel = modelLOCAL,
 verbose = TRUE,
 pvalue.hist = FALSE);


# Pull results - t-statistic and p-value
beta = me$all$eqtls$beta;
tstat = me$all$eqtls$statistic;
pvalue = me$all$eqtls$pvalue;
rez = c(beta = beta, tstat = tstat, pvalue = pvalue);
print(rez)

# Results from linear regression
lmdl = lm( gene.mat ~ snps.mat + cvrt.mat + local.mat);
lmout = summary(lmdl)$coefficients[2,c("Estimate","t value","Pr(>|t|)")];
print( lmout );
}
